refactor: update date imports and function names and locations
